Moved NewTexture to global, instead of OS-specific
