+++ assets.go
@@ -274,3 +274,25 @@ func LoadShader(vertSrc, fragSrc string) *gl.Program {
 
 	return program
 }
+
+func NewTexture(img Image) *Texture {
+	var id *gl.Texture
+	if !headless {
+		id = Gl.CreateTexture()
+
+		Gl.BindTexture(Gl.TEXTURE_2D, id)
+
+		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_WRAP_S, Gl.CLAMP_TO_EDGE)
+		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_WRAP_T, Gl.CLAMP_TO_EDGE)
+		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_MIN_FILTER, Gl.LINEAR)
+		Gl.TexParameteri(Gl.TEXTURE_2D, Gl.TEXTURE_MAG_FILTER, Gl.NEAREST)
+
+		if img.Data() == nil {
+			panic("Texture image data is nil.")
+		}
+
+		Gl.TexImage2D(Gl.TEXTURE_2D, 0, Gl.RGBA, Gl.RGBA, Gl.UNSIGNED_BYTE, img.Data())
+	}
+
+	return &Texture{id, float32(img.Width()), float32(img.Height())}
+}
